Return u32 from entity-id and source-sn accessors (#13)
Base zenoh's `EntityId` and `SourceSn` are both `u32`. The flat accessors
returned `i32`/`i64` via `as` casts, which misrepresents entity ids ≥ 2^31 and
is inconsistent with flat's own `EntityGlobalId.eid` (already `u32`). Return
`u32` directly from:

- sample_get_source_eid, sample_get_source_sn
- publisher_get_eid, subscriber_get_eid
- querier_get_eid, queryable_get_eid
- reply_get_replier_eid

Absence handling (currently `unwrap_or(0)`) is unchanged here — that is tracked
separately in #10.

+++ src/base/publisher/mod.rs
@@ -66,6 +66,6 @@ pub fn publisher_get_zid(publisher: &Publisher) -> ZenohId {
 /// This information is available only when unstable features are enabled.
 #[cfg(feature = "unstable")]
 #[prebindgen(cfg = "feature = \"unstable\"")]
-pub fn publisher_get_eid(publisher: &Publisher) -> i32 {
-    publisher.id().eid() as i32
+pub fn publisher_get_eid(publisher: &Publisher) -> u32 {
+    publisher.id().eid()
 }
